fix(contract): remove unnecessary u32 cast in deposit_batch
Vec: :len() already returns u32 in Soroban, so clippy rejects items.len() as u32 under -D warnings.

+++ Dechat/stellar-contracts/src/lib.rs
@@ -1598,7 +1598,7 @@ impl FiatBridge {
 
         Self::require_not_paused(&env)?;
 
-        let total_ops = items.len() as u32;
+        let total_ops = items.len();
         let current_ledger = env.ledger().sequence();
 
         // Read shared config once
